refactor(storage): 🔧 Simplify local storage configuration and improve code clarity
