+++ apps/docs/content/docs/dev/storage/local.mdx
@@ -78,12 +78,6 @@ import { TypeTable } from "fumadocs-ui/components/type-table";
 
 <TypeTable
   type={{
-    uploadsDir: {
-      description:
-        "Directory (relative to the working directory) files are written to.",
-      type: "string",
-      default: "public/uploads",
-    },
     publicPath: {
       description: "URL path prefix the files are served under.",
       type: "string",
